# 2/hw2/hw2/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
...
  /test/get:
    # binds a127 app logic to a route
    x-swagger-router-controller: test
    x-a127-apply: {}
    get:
      description: Returns info about request
      # used as the method name of the controller
      operationId: testGet
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    # handle other requests that may come in
    put: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    post: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    delete: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
  /test/post:
    # binds a127 app logic to a route
    x-swagger-router-controller: test
    x-a127-apply: {}
    post:
      description: Returns info about request
      # used as the method name of the controller
      operationId: testPost
      parameters:
        - name: item
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    # handle other requests that may come in
    put: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    get: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    delete: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
  /test/put:
    # binds a127 app logic to a route
    x-swagger-router-controller: test
    x-a127-apply: {}
    put:
      description: Returns info about request
      # used as the method name of the controller
      operationId: testPut
      parameters:
        - name: item
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    # handle other requests that may come in
    get: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    post: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    delete: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
  /test/delete:
    # binds a127 app logic to a route
    x-swagger-router-controller: test
    x-a127-apply: {}
    delete:
      description: Returns info about request
      # used as the method name of the controller
      operationId: testDelete
      parameters:
        - name: item
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    # handle other requests that may come in
    put: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    post: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    get: 
      operationId: error
      responses: 
        "405":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string